athenaos-ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
<<<<<<< HEAD
transformers==4.41.2
torch==2.3.0
sentence-transformers==2.7.0
httpx==0.27.0
=======
httpx==0.27.0
google-generativeai==0.5.4

>>>>>>> bedd0c9e5da5c5b45f374247e464d676e0550a32
